Made PI and RAD_TO_DEG  long doubles
** PI and RAD_TO_DEG  long doubles from doubles

+++ math.h
@@ -1,7 +1,7 @@
 #pragma once; 
 namespace Skele_lib{
 	namespace Math{
-		const double PI = 3.14159265358979323846;
-		const double RAD_TO_DEG = 180 / PI; 
+		const long double PI = 3.14159265358979323846;
+		const long double RAD_TO_DEG = 180 / PI; 
 	}
 }
